GLOB-38241 support string values for memcached timeout
Part of [GLOB-38241]

We'd like to be able to reduce the memcached `timeout` (if it's failing
then we skip the cache and just rerequest the value we need). This value
comes in as an environment variable as a string, but Node's `Stream`
blows up as it expects a number.

This commit:

* makes sure the `timeout` is numeric if it exists
* corrects the constructor signature in the docblock
* adds a couple more tests